+++ FocusLit-Server/src/main/java/com/focus/lit/service/impl/JwtServiceImpl.java
@@ -82,7 +82,7 @@ public class JwtServiceImpl {
         return Keys.hmacShaKeyFor(keyBytes);
     }
 
-    public void deleteUserTokens(Long userId){
+    public void deleteUserTokens(int userId){
         tokenRepository.deleteUserTokens(userId);
     }
 }
